churnguard-backend/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
bcrypt==5.0.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
deprecation==2.1.0
ecdsa==0.19.2
fastapi==0.135.3
flatbuffers==25.12.19
fsspec==2026.3.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
imbalanced-learn==0.14.1
joblib==1.5.3
kaggle==2.0.1
keras==3.14.0
libclang==18.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
multidict==6.7.1
namex==0.1.0
numpy==2.4.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.2
passlib==1.7.4
postgrest==2.28.3
propcache==0.4.1
protobuf==7.34.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.26
realtime==2.28.3
requests==2.33.1
rich==14.3.4
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sklearn-compat==0.1.5
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
websockets==15.0.1
wrapt==2.1.2
xgboost==3.2.0
yarl==1.23.0
zstandard==0.25.0
